Fix overflow for threshold in X+Y=Z

+++ solver/src/main/java/org/chocosolver/solver/constraints/ternary/PropXplusYeqZ.java
@@ -89,7 +89,7 @@ public class PropXplusYeqZ extends Propagator<IntVar> {
         int lb = vars[v1].getLB() + vars[v2].getLB();
         int ub = vars[v1].getUB() + vars[v2].getUB();
         boolean change = vars[vr].updateBounds(lb, ub, this);
-        if (vars[v1].getDomainSize() * vars[v2].getDomainSize() > THRESHOLD) return change;
+        if ((long) vars[v1].getDomainSize() * vars[v2].getDomainSize() > THRESHOLD) return change;
         if (!allbounded) {
             set.clear();
             int ub1 = vars[v1].getUB();
@@ -125,7 +125,7 @@ public class PropXplusYeqZ extends Propagator<IntVar> {
         int lb = vars[v1].getLB() - vars[v2].getUB();
         int ub = vars[v1].getUB() - vars[v2].getLB();
         boolean change = vars[vr].updateBounds(lb, ub, this);
-        if (vars[v1].getDomainSize() * vars[v2].getDomainSize() > THRESHOLD) return change;
+        if ((long) vars[v1].getDomainSize() * vars[v2].getDomainSize() > THRESHOLD) return change;
         if (!allbounded) {
             set.clear();
             int ub1 = vars[v1].getUB();
